All foodsec indicators

All foodsec indicators

Usage

foodsec(
  .tbl,
  fcs_cereal,
  fcs_legumes,
  fcs_dairy,
  fcs_meat,
  fcs_veg,
  fcs_fruit,
  fcs_oil,
  fcs_sugar,
  fcs_cat = "normal",
  hhs_lev1_nofoodhh,
  hhs_lev2_nofoodhh,
  hhs_lev1_sleephungry,
  hhs_lev2_sleephungry,
  hhs_lev1_alldaynight,
  hhs_lev2_alldaynight,
  hhs_level1_codes,
  hhs_level2_codes,
  rcsi_lesspreferred,
  rcsi_borrowfood,
  rcsi_limitportion,
  rcsi_restrict,
  rcsi_reducemeals,
  lcs_stress_1,
  lcs_stress_2,
  lcs_stress_3,
  lcs_stress_4,
  lcs_crisis_1,
  lcs_crisis_2,
  lcs_crisis_3,
  lcs_emergency_1,
  lcs_emergency_2,
  lcs_emergency_3,
  lcs_level_codes
)

Arguments

.tbl

Data

fcs_cereal

Cereals component column

fcs_legumes

Legumes component column

fcs_dairy

Dairy component column

fcs_meat

Meat component column

fcs_veg

Vegetables component column

fcs_fruit

Fruit component column

fcs_oil

Oil component column

fcs_sugar

Sugar component column

fcs_cat

Cut-offs categories, either "normal" (21.5/35) or "alternate" (28/42)

hhs_lev1_nofoodhh

Component column: No food of any kind in the house

hhs_lev2_nofoodhh

Follow-up frequency column

hhs_lev1_sleephungry

Component column: Go to sleep hungry because there was not enough food

hhs_lev2_sleephungry

Follow-up frequency column

hhs_lev1_alldaynight

Component column: Go a whole day and night without eating

hhs_lev2_alldaynight

Follow-up frequency column

hhs_level1_codes

Character vector of at least "Yes" and "No" codes (in this order), e.g. c("yes", "no")

hhs_level2_codes

Character vector of at least frequencies codes, in the following order: "Rarely", "Sometimes", "Often", e.g. c("rarely", "sometimes", "often")

rcsi_lesspreferred

Component column: Rely on less preferred and less expensive food

rcsi_borrowfood

Component column: Borrow food or rely on help from friends or relatives

rcsi_limitportion

Component column: Limit portion size at mealtime

rcsi_restrict

Component column: Restrict consumption by adults in order for small children to eat

rcsi_reducemeals

Component column: Reduce the number of meals eaten in a day

lcs_stress_1

Component column: Stress strategy 1

lcs_stress_2

Component column: Stress strategy 2

lcs_stress_3

Component column: Stress strategy 3

lcs_stress_4

Component column: Stress strategy 4

lcs_crisis_1

Component column: Crisis strategy 1

lcs_crisis_2

Component column: Crisis strategy 2

lcs_crisis_3

Component column: Crisis strategy 3

lcs_emergency_1

Component column: Emergency strategy 1

lcs_emergency_2

Component column: Emergency strategy 2

lcs_emergency_3

Component column: Emergency strategy 3

lcs_level_codes

Character vector of at responses codes, in the following order: ""Yes", "No, exhausted", "No, no need", "No, not applicable", e.g. c("yes", "exhausted", "no_need", "not_applicable")

Value

All columns produced by the following impactR functions: fcs(), hhs(), rcsi(), lcs(), fcm(), fcp() and fclp().