Predict mechanical loading

Predict either ground reaction force or loading rate, or both, based on accelerometer data.

predict_loading(data, outcome, vector, model)

Arguments

data

An impactr_data object, as obtained with read_acc().

outcome

A character string. Can be either "grf" (for ground reaction force), or "lr" (for loading rate) or "all" (for both mechanical loading variables).

vector

A character string indicating in which acceleration vector to find the peaks. Can be "resultant", "vertical" or "all".

model

A character string indicating which model to use to make the predictions. The values currently supported are "walking", "walking/running" and "jumping".

Value

An object of class impactr_peaks with the ground reaction force and/or loading rate peaks magnitude stored in the columns.

Examples

data <- read_acc(impactr_example("hip-raw.csv"))
data <- specify_parameters(data, acc_placement = "hip", subj_body_mass = 78)
data <- find_peaks(data, vector = "vertical")
predict_loading(
  data,
  outcome = "grf",
  vector = "vertical",
  model = "walking/running"
)
#> # Start time:              2021-04-06 15:43:00
#> # Sampling frequency:      100Hz
#> # Accelerometer placement: Hip
#> # Subject body mass:       78kg
#> # Filter:                  No filter applied
#> # Data dimensions:         251 × 3
#>    timestamp           vertical_peak_acc vertical_peak_grf
#>    <dttm>                          <dbl>             <dbl>
#>  1 2021-04-06 15:43:00              1.83             1484.
#>  2 2021-04-06 15:43:03              1.41             1433.
#>  3 2021-04-06 15:43:04              1.59             1455.
#>  4 2021-04-06 15:43:06              1.35             1426.
#>  5 2021-04-06 15:43:09              2.61             1579.
#>  6 2021-04-06 15:43:11              1.38             1430.
#>  7 2021-04-06 15:43:14              1.42             1435.
#>  8 2021-04-06 15:43:16              1.36             1428.
#>  9 2021-04-06 15:43:16              1.46             1440.
#> 10 2021-04-06 15:43:17              1.32             1423.
#> # … with 241 more rows