Specify prediction model parameters

Specify the accelerometer placement used and the subject body mass. These data is needed in order to use the mechanical loading prediction models.

specify_parameters(data, acc_placement, subj_body_mass)

Arguments

data
An impactr_data object, as obtained with read_acc().

acc_placement
A character string indicating the accelerometer placement. Can be either "ankle", "back", or "hip".

subj_body_mass
A double scalar indicating the subject body mass in kilograms.

Value

An object of class impactr_data with the specified parameters as attributes.

Examples

data <- read_acc(impactr_example("hip-raw.csv"))
specify_parameters(data, acc_placement = "hip", subj_body_mass = 79.2)
#> # Start time:              2021-04-06 15:43:00
#> # Sampling frequency:      100Hz
#> # Accelerometer placement: Hip
#> # Subject body mass:       79.2kg
#> # Filter:                  No filter applied
#> # Data dimensions:         30,000 × 4
#>    timestamp           acc_X  acc_Y acc_Z
#>    <dttm>              <dbl>  <dbl> <dbl>
#>  1 2021-04-06 15:43:00 0.262 -0.688 0.063
#>  2 2021-04-06 15:43:00 0.25  -0.727 0.039
#>  3 2021-04-06 15:43:00 0.254 -0.816 0.191
#>  4 2021-04-06 15:43:00 0.258 -0.891 0.367
#>  5 2021-04-06 15:43:00 0.281 -0.914 0.344
#>  6 2021-04-06 15:43:00 0.316 -0.922 0.23 
#>  7 2021-04-06 15:43:00 0.32  -0.891 0.203
#>  8 2021-04-06 15:43:00 0.332 -0.926 0.109
#>  9 2021-04-06 15:43:00 0.363 -1.02  0.168
#> 10 2021-04-06 15:43:00 0.418 -0.996 0.387
#> # … with 29,990 more rows