Summarise accelerometer non-wear time

Validates each measurement day based on the minimum number of wear hours and validates the entire observation based on the minimum number of valid days. It updates the values of the wear column (created by mark_nonwear()) of the impactr_data object provided by data, marking as non-wear time the days considered to be invalid. It also creates a wear time daily summary that can be saved in a csv file.

Usage

summarise_nonwear(data, min_hour_crit, min_day_crit, save_summary)

Arguments

data

An impactr_data object with the wear column, as obtained with mark_nonwear().

min_hour_crit

The minimum number of hours marked as wear time in a day for it to be considered valid.

min_day_crit

The minimum number of valid days for the data of a given subject to be considered valid.

save_summary

Indicates whether of not to save the wear time daily summary to a csv file. Provide a valid path to a file ending with the ".csv" extension as a character string if you want the summary to be saved.

Value

The impactr_data object provided by data with the wear column values updated based on the validations. If no data are classified as valid, it returns NULL. Also saves the summary in a csv file if save_summary is a path.

Details

This is an internal function, designed to be used inside the wrapper function remove_nonwear() that integrates all steps of detection, removal and inspection of accelerometer non-wear time.

See also