Get path to example data

impactr comes with some example ActiGraph accelerometer raw data files in its inst/extdata directory. This function make them easy to access.

impactr_example(file = NULL)

Arguments

file
A character string with the file name. If NULL, the example files will be listed.

Value

If file = NULL, it returns the file names of the example data files, else it returns the path to the example data.

Examples

impactr_example()
#> [1] "hip-raw.csv"
impactr_example("hip-raw.csv")
#> [1] "/Users/runner/work/_temp/Library/impactr/extdata/hip-raw.csv"